wetlab/protein-to-vector/requirements.txt
dnachisel
biopython